+++ sydtest/CHANGELOG.md
@@ -5,11 +5,14 @@
 ### Changed
 
 * The `TestDefM` now contains a `TestDefEnv` which also contains the test description path, along with `TestRunSettings`.
-  It's unlikely that you depended on this, but requires a major version bump nonetheless.
+* Removed the `MonadState ()` instance of `TestDefM`. It was just silly.
+* Changed the internals of `TestDefM` to use `ReaderT` and a strict `WriterT` instead of `RWST`.
+* Renamed `wrapRWST` to `wrapForest`.
 
 ### Added
 
 * `getTestDescriptionPath` to get the test description path upwards from inside a test definition.
+
 
 ## [0.7.0.1] - 2021-12-23